Change migration from float to double for latitude and longitude

+++ database/migrations/2015_07_01_202306_create_locations_table.php
@@ -16,8 +16,8 @@ class CreateLocationsTable extends Migration
             $table->increments('id');
             $table->string('street')->nullable();
             $table->string('postcode')->nullable();
-            $table->float('latitude')->nullable();
-            $table->float('longitude')->nullable();
+            $table->double('latitude')->nullable();
+            $table->double('longitude')->nullable();
             $table->date('lastDate')->nullable();
             $table->integer('city_id')->nullable();
             $table->integer('user_id')->unsigned();
